Change utils to be modular
